homey-api/requirements.txt
docker==5.0.3
Flask==2.0.2
Flask_Cors==3.0.10
python-dotenv==0.19.2
python_dateutil==2.8.2
PyYAML < 6.0.0
requests==2.25.1
urllib3==1.26.5
Werkzeug==2.0.2
gunicorn==20.1.0
requests-futures==1.0.0